+++ doc/release-notes.md
@@ -5,6 +5,7 @@
 - kiali connection to tempo tracing added, spring application name now contains the Namespace
 - upgrade to apisix chart 2.8.0
 - upgrade to kind kubernetes 1.3.0
+- upgrade to kubernetes dashboard 6.0.8 (7.x has a lot of breaking changes)
 
 # 1.3.1
 - switchable kafka and rabbitmq broker added